Release/v1.0.0 (#57)
* Fixing issues of not working properly under postgres database; Various fixes around the entire app - mostly comments and typos

* bump version of all workspace crates

* removing chatgpt out of the readme :)

* remove println and update readme

* Adding a release version checker

* adding logic in the publish job to extract the tag number for docker tag

* changing the test action trigger

* changing the trigger again

* fixing title name typo

* fixing preview issues

* Fixing some typos and wording in the readme

+++ fs/src/contract.rs
@@ -16,31 +16,31 @@ pub trait FsProviderContract {
     async fn write<T: IntoFilename>(&self, filename: &T, data: &[u8]) -> AppResult<()>;
 
     /// Check if the chunk already exists in the storage provider
-    async fn exists<T: IntoFilename>(&self, filename: &T, chunk: i32) -> AppResult<bool>;
+    async fn exists<T: IntoFilename>(&self, filename: &T, chunk: i64) -> AppResult<bool>;
 
     /// Get a file representation from the storage provider
-    async fn get<T: IntoFilename>(&self, filename: &T, chunk: i32) -> AppResult<File>;
+    async fn get<T: IntoFilename>(&self, filename: &T, chunk: i64) -> AppResult<File>;
 
     /// Get a file representation from the storage provider
     async fn all<T: IntoFilename>(&self, filename: &T) -> AppResult<Vec<File>>;
 
     /// Push specific data chunk into a part file
-    async fn push<T: IntoFilename>(&self, filename: &T, chunk: i32, data: &[u8]) -> AppResult<()>;
+    async fn push<T: IntoFilename>(&self, filename: &T, chunk: i64, data: &[u8]) -> AppResult<()>;
 
     /// Pull data chunk of a file from the storage provider.
-    async fn pull<T: IntoFilename>(&self, filename: &T, chunk: i32) -> AppResult<Vec<u8>>;
+    async fn pull<T: IntoFilename>(&self, filename: &T, chunk: i64) -> AppResult<Vec<u8>>;
 
     /// Purge all the parts for a file from the storage provider.
     async fn purge<T: IntoFilename>(&self, filename: &T) -> AppResult<()>;
 
     /// Get a vector of chunk indexes that were already uploaded so we can resume
     /// the upload process on the frontend without doing the double work.
-    async fn get_uploaded_chunks<T: IntoFilename>(&self, filename: &T) -> AppResult<Vec<i32>>;
+    async fn get_uploaded_chunks<T: IntoFilename>(&self, filename: &T) -> AppResult<Vec<i64>>;
 
     /// Return stream of either one file chunk, or all chunks if no file chunk is specified.
     async fn stream<T: IntoFilename>(
         &self,
         filename: &T,
-        chunk: Option<i32>,
+        chunk: Option<i64>,
     ) -> AppResult<Streamer>;
 }
